removed updated_at user field and removed superuser flag from apis

+++ QuizApi/authentication/views.py
@@ -54,6 +54,8 @@ class AdminRetrieveUpdateAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
+        if(request.user and not request.user.is_superuser):
+            return Response({"detail": "You do not have permission to perform this action."},status=status.HTTP_403_FORBIDDEN)
         user = self.get_object(pk)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
